Extracts helper functions to utils.js
